moved books into its own modules
