+++ tests/unit/store/actions.spec.js
@@ -28,32 +28,6 @@ describe("actions", () => {
     });
   });
 
-  describe("loadBooks", () => {
-    beforeEach(() => {
-      httpService.get.mockImplementation((_, callback) =>
-        callback(200, { data: [{ title: "Go Blue" }] })
-      );
-      actions.loadBooks({ commit });
-    });
-
-    test("loadBooks calls commit with START_LOADING", () => {
-      expect(commit).toHaveBeenCalledWith("START_LOADING");
-    });
-    test("loadBooks calls commit with ADD_API_REQUEST", () => {
-      expect(commit).toHaveBeenCalledWith("ADD_API_REQUEST", {
-        url: "GET /api/v1/books",
-        response: { data: [{ title: "Go Blue" }] },
-        status: 200,
-      });
-    });
-    test("loadBooks calls commit with SET_BOOKS", () => {
-      expect(commit).toHaveBeenCalledWith("SET_BOOKS", [{ title: "Go Blue" }]);
-    });
-    test("loadBooks calls commit with STOP_LOADING", () => {
-      expect(commit).toHaveBeenCalledWith("STOP_LOADING");
-    });
-  });
-
   describe("clearRepos", () => {
     test("initialize calls commit with SET_REPOS", () => {
       actions.clearRepos({ commit });
